@@ -1,8 +1,9 @@
 package org.anystub;
 
-import org.junit.Test;
 
-import static org.junit.Assert.*;
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 @AnyStubId
 public class AnyStubFileLocatorTest {
