@@ -1,13 +1,10 @@
 package org.anystub;
 
 import org.anystub.mgmt.BaseManagerFactory;
-import org.junit.Ignore;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
-import static org.junit.Assert.assertEquals;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.verify;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.mockito.Mockito.*;
 
 public class BaseMockTest {
 
@@ -16,7 +13,7 @@ public class BaseMockTest {
     }
 
     @Test
-    public void passThroughTest() throws Exception {
+    public void testPassThroughTest() throws Exception {
         MockSupplier upstream = mock(MockSupplier.class);
 
 
