@@ -1,25 +1,24 @@
 package org.anystub.it;
 
 import org.anystub.Base;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
+import org.springframework.boot.test.context.TestConfiguration;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
 
 import java.io.IOException;
 
-import static org.junit.Assert.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
 
 /**
  * Created
  */
 
-@RunWith(SpringRunner.class)
 @SpringBootTest
 public class WorkerTest {
-
-
 
     @Autowired
     private Worker worker;
@@ -31,12 +30,12 @@ public class WorkerTest {
     private Base base;
 
     @Test
-    public void xTest() throws IOException {
+    public void testXTest() throws IOException {
         assertEquals("fixed", sourceSystem.get());
     }
 
     @Test
-    public void xxTest() throws IOException {
+    public void testXxTest() throws IOException {
         base.clear();
 
         assertEquals("fixed", worker.get());
@@ -45,12 +44,13 @@ public class WorkerTest {
     }
 
     @Test
-    public void randTest() throws IOException {
+    public void testRandTest() throws IOException {
         base.clear();
-        assertEquals("1", -1594594225, worker.rand());
-        assertEquals("2", -1594594225, worker.rand());
+        assertEquals(-1594594225, worker.rand());
+        assertEquals(-1594594225, worker.rand());
         assertEquals(2L, base.times());
 
     }
+
 
 }
