@@ -1,42 +1,38 @@
 package org.anystub.it_http;
 
 import org.anystub.AnyStubId;
-import org.anystub.Base;
 import org.anystub.http.AnySettingsHttp;
-import org.apache.http.client.HttpClient;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.http.HttpEntity;
+import org.springframework.http.HttpMethod;
 import org.springframework.http.ResponseEntity;
-import org.springframework.test.context.junit4.SpringRunner;
+import org.springframework.util.LinkedMultiValueMap;
+import org.springframework.util.MultiValueMap;
 import org.springframework.web.client.HttpClientErrorException;
 import org.springframework.web.client.RestTemplate;
 
+import java.util.Arrays;
 
 import static org.anystub.mgmt.BaseManagerFactory.getStub;
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertTrue;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
-@RunWith(SpringRunner.class)
-@SpringBootTest()
+@SpringBootTest
 @AnyStubId(filename = "httpStub.yml")
 public class HttpSourceSystemTest {
 
     @Autowired
     private HttpSourceSystem httpSourceSystem;
 
-    @Autowired
-    private Base httpBase;
 
     @Autowired(required = false)
     private RestTemplate restTemplate;
 
-    @Autowired
-    HttpClient httpClient;
-
     @Test
-    public void getStringsTest() {
+    public void testGetStrings() {
 
         String r = httpSourceSystem.getStrings();
 
@@ -45,22 +41,35 @@ public class HttpSourceSystemTest {
     }
 
     @Test
-    @AnyStubId
+    @AnyStubId(filename = "getWithHeaders")
     @AnySettingsHttp(allHeaders = true)
-    public void getWithHeaders() {
+    public void testGetWithHeaders() {
 
-        ResponseEntity<String> forEntity = restTemplate.getForEntity("https://gturnquist-quoters.cfapps.io/api/random", String.class);
+        MultiValueMap<String, String> headers = new LinkedMultiValueMap<>();
+        headers.add("Content-Type", "application/json");
+        headers.add("Authorization", "tokenxxx");
+        ResponseEntity<String> forEntity = restTemplate.exchange(
+                "https://gturnquist-quoters.cfapps.io/api/random", HttpMethod.GET, new HttpEntity<Object>(headers),
+                String.class);
+
         assertEquals(200, forEntity.getStatusCodeValue());
 
         assertEquals(1, getStub().times());
-        assertTrue(getStub().match().findFirst().get().getKey(2).startsWith("Accept:"));
+        String [] v= new String[3];
+        v[0] = getStub().match().findFirst().get().getKey(2);
+        v[1] = getStub().match().findFirst().get().getKey(3);
+        v[2] = getStub().match().findFirst().get().getKey(4);
+        Arrays.sort(v);
+        assertTrue(v[0].startsWith("Accept:"));
+        assertTrue(v[1].startsWith("Authorization:"));
+        assertTrue(v[2].startsWith("Content-Type:"));
 
     }
 
     @Test
-    @AnyStubId
+    @AnyStubId(filename = "getWithHeader")
     @AnySettingsHttp(headers = "Accept")
-    public void getWithHeader() {
+    public void testGetWithHeader() {
 
         ResponseEntity<String> forEntity = restTemplate.getForEntity("https://gturnquist-quoters.cfapps.io/api/random", String.class);
         assertEquals(200, forEntity.getStatusCodeValue());
@@ -70,66 +79,58 @@ public class HttpSourceSystemTest {
 
     }
 
-    @Test(expected = HttpClientErrorException.class)
-    public void postTest() {
-        restTemplate.postForEntity("https://gturnquist-quoters.cfapps.io/api/random", null, String.class);
-    }
-
-    @Test(expected = HttpClientErrorException.class)
-    @AnySettingsHttp(bodyTrigger = "random/xxx")
-    public void postBodyTest() {
-        restTemplate.postForEntity("https://gturnquist-quoters.cfapps.io/api/random/xxx", "{test}", String.class);
+    @Test
+    public void testPost() {
+        Assertions.assertThrows(HttpClientErrorException.class, () -> {
+            restTemplate.postForEntity("https://gturnquist-quoters.cfapps.io/api/random", null, String.class);
+        });
     }
 
     @Test
-    @AnyStubId
     @AnySettingsHttp(bodyTrigger = "random/xxx")
-    public void postMultiLineBodyTest() {
-        try {
+    public void testPostBody() {
+        Assertions.assertThrows(HttpClientErrorException.class, () -> {
+            restTemplate.postForEntity("https://gturnquist-quoters.cfapps.io/api/random/xxx", "{test}", String.class);
+        });
+
+        assertEquals(1, getStub().times(null, null, null, "{test}"));
+    }
+
+    @Test
+    @AnyStubId(filename = "postMultiLineBodyTest")
+    @AnySettingsHttp(bodyTrigger = "random/xxx")
+    public void testPostMultiLineBody() {
+        Assertions.assertThrows(HttpClientErrorException.class, () -> {
             restTemplate.postForEntity("https://gturnquist-quoters.cfapps.io/api/random/xxx", "{test:1,\ntest2:3}", String.class);
-        } catch (HttpClientErrorException ex) {
+        });
 
-        }
-        try {
-
+        Assertions.assertThrows(HttpClientErrorException.class, () -> {
             restTemplate.postForEntity("https://gturnquist-quoters.cfapps.io/api/random/xxx", "{test:1,\n\ntest2:3}", String.class);
-        } catch (HttpClientErrorException ex) {
-
-        }
+        });
 
         assertEquals(2, getStub().times());
     }
 
     @Test
-    @AnyStubId
+    @AnyStubId(filename = "postBinaryBodyTest")
     @AnySettingsHttp(bodyTrigger = "random/xxx")
-    public void postBinaryBodyTest() {
-        try {
-            restTemplate.postForEntity("https://gturnquist-quoters.cfapps.io/api/random/xxx", new String(new byte[]{1, 2, 3, 4}), String.class);
-        } catch (HttpClientErrorException ex) {
-
-        }
+    public void testPostBinaryBody() {
+        restTemplate.postForEntity("https://gturnquist-quoters.cfapps.io/api/random/xxx", new String(new byte[]{1, 2, 3, 4}), String.class);
 
         assertEquals(1, getStub().matchEx("POST", null, null, "BASE64.*").count());
     }
 
     @Test
-    @AnyStubId
+    @AnyStubId(filename = "postMaskedBodyTest")
     @AnySettingsHttp(bodyTrigger = "random/xxx", bodyMask = "\\{.+\\}")
-    public void postMaskedBodyTest() {
-        int exceptionsNumber = 0;
-
+    public void testPostMaskedBody() {
 
         for (int i = 0; i < 2; i++) {
-
-            try {
-
-                restTemplate.postForEntity("https://gturnquist-quoters.cfapps.io/api/random/xxx", "some text{" + i + "} suffix", String.class);
-            } catch (HttpClientErrorException ex) {
-                exceptionsNumber++;
-            }
+            int callNumber = i;
+            Assertions.assertThrows(HttpClientErrorException.class, ()-> {
+                restTemplate.postForEntity("https://gturnquist-quoters.cfapps.io/api/random/xxx", "some text{" + callNumber + "} suffix", String.class);
+            });
         }
-        assertEquals(2, exceptionsNumber);
 
         assertEquals(2,
                 getStub()
@@ -140,7 +141,7 @@ public class HttpSourceSystemTest {
     }
 
     @Test
-    public void getBase64Test() {
+    public void testResponseBody() {
         ResponseEntity<String> forEntity = restTemplate.getForEntity("https://test", String.class);
         assertEquals("test", forEntity.getBody());
     }
