@@ -1,14 +1,16 @@
 package org.anystub.jdbc;
 
-import org.junit.Test;
+
+import org.junit.jupiter.api.Test;
 
 import java.util.Iterator;
 
-import static org.junit.Assert.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
 
 public class EncoderSavepointTest {
     @Test
-    public void encodeTest() {
+    public void testEncodeTest() {
         StubSavepoint stubSavepoint = new StubSavepoint(1, "2");
         Iterator<String> encode = new EncoderSavepoint().encode(stubSavepoint).iterator();
         assertEquals("Savepoint",encode.next());
