@@ -1,16 +1,16 @@
 package org.anystub.mgmt;
 
 import org.anystub.Base;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 import java.io.File;
 
-import static org.junit.Assert.assertTrue;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 public class BaseManagerImplTest {
 
     @Test
-    public void scenario() {
+    public void testScenario() {
 
         BaseManagerImpl.instance().getBase("test.yml");
 
@@ -25,7 +25,7 @@ public class BaseManagerImplTest {
     }
 
     @Test
-    public void getNamesTest() {
+    public void testGetNamesTest() {
         assertTrue(BaseManagerImpl.getFilePath("test3.yml").endsWith(File.separator + "test3.yml"));
         assertTrue(BaseManagerImpl.getFilePath("src/test3.yml").endsWith(File.separator + "test3.yml"));
         assertTrue(BaseManagerImpl.getFilePath("./test3.yml").endsWith(File.separator + "test3.yml"));
