@@ -1,9 +1,9 @@
-package org.anystub;
+package org.anystub.it;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
-@SpringBootApplication
+@SpringBootApplication()
 public class AnystubApplication {
 
 	public static void main(String[] args) {
