@@ -1,31 +1,23 @@
 package org.anystub.it;
 
 import org.anystub.Base;
-import org.anystub.http.StubHttpClient;
-import org.anystub.jdbc.StubDataSource;
 import org.anystub.mgmt.BaseManagerFactory;
-import org.apache.http.client.HttpClient;
-import org.apache.http.impl.client.HttpClientBuilder;
-import org.h2.jdbcx.JdbcDataSource;
+import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.http.client.ClientHttpRequestFactory;
-import org.springframework.http.client.HttpComponentsClientHttpRequestFactory;
-import org.springframework.web.client.RestTemplate;
+import org.springframework.context.annotation.Primary;
+import org.springframework.test.context.ContextConfiguration;
 
-import javax.sql.DataSource;
 import java.io.IOException;
 
-/**
- *
- */
 @Configuration
 public class DefaultConfiguration {
-
     @Bean
     public String externalSystemUrl() {
         return "http://localhost:8080";
     }
+
+
 
     @Bean
     Base base() {
@@ -35,9 +27,8 @@ public class DefaultConfiguration {
     }
 
     @Bean
-    public SourceSystem sourceSystem(Base base) {
-
-
+    @Primary
+    public SourceSystem sourceSystemTest(Base base) {
         return new SourceSystem("http://localhost:8080") {
             @Override
             public String get() throws IOException {
@@ -58,32 +49,4 @@ public class DefaultConfiguration {
         };
     }
 
-    @Bean
-    public HttpClient createHttpClient(Base httpBase) {
-
-        HttpClient real = HttpClientBuilder.create().build();
-        return StubHttpClient.stub(real).setFallbackBase(httpBase);
-
-    }
-
-    @Bean
-    public ClientHttpRequestFactory requestFactory(HttpClient httpClient) {
-        HttpComponentsClientHttpRequestFactory requestFactory = new HttpComponentsClientHttpRequestFactory();
-        requestFactory.setHttpClient(httpClient);
-        return requestFactory;
-    }
-
-    @Bean
-    public RestTemplate restTemplate(ClientHttpRequestFactory requestFactory) {
-
-        return new RestTemplate(requestFactory);
-    }
-
-    @Bean
-    public DataSource dataSource() {
-
-        JdbcDataSource ds = new JdbcDataSource();
-        ds.setURL("jdbc:h2:./test3;DB_CLOSE_ON_EXIT=FALSE;AUTO_RECONNECT=TRUE");
-        return StubDataSource.stub(ds);
-    }
 }
